Delete replaced tests
